pyproject.toml
[tool.poetry]
name = "aqi-watcher"
version = "0.1.0"
description = ""
authors = ["Girish Bhatta <girishbhatta93@gmail.com>"]
readme = "README.md"
packages = [{include = "aqi_watcher", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
apache-airflow = "2.7.3"
asyncio = "^3.4.3"
aiohttp = "^3.9.1"
black = "^23.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
